Introduce base/storage to represent the Web Storage API and persistence-related customizations
