+++ react/features/base/redux/index.js
@@ -1,6 +1,3 @@
 export * from './functions';
 export { default as MiddlewareRegistry } from './MiddlewareRegistry';
-export { default as PersistencyRegistry } from './PersistencyRegistry';
 export { default as ReducerRegistry } from './ReducerRegistry';
-
-import './middleware';
