+++ react/features/base/storage/_.native.js
@@ -1,0 +1,1 @@
+export * from './native';
