+++ react/features/base/storage/index.js
@@ -1,0 +1,4 @@
+export * from './_';
+export { default as PersistenceRegistry } from './PersistenceRegistry';
+
+import './middleware';
