+++ react/features/base/storage/native/index.js
@@ -1,0 +1,1 @@
+import './polyfills-browser';
